test(dataprotection): align restore test filenames
