+++ pkg/dataprotection/restore/manager_test.go
@@ -20,20 +20,25 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 package restore
 
 import (
+	"context"
 	"fmt"
 	"strconv"
 	"strings"
+	"testing"
 	"time"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
+	"github.com/stretchr/testify/require"
 	batchv1 "k8s.io/api/batch/v1"
 	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
+	"sigs.k8s.io/controller-runtime/pkg/client/fake"
 
 	appsv1 "github.com/apecloud/kubeblocks/apis/apps/v1"
 	dpv1alpha1 "github.com/apecloud/kubeblocks/apis/dataprotection/v1alpha1"
@@ -57,6 +62,71 @@ func postReadyJobEnvValues(job *batchv1.Job, name string) []string {
 		}
 	}
 	return values
+}
+
+func TestRestoreManagerPostReadyTargetEnv(t *testing.T) {
+	const (
+		namespace     = "default"
+		clusterName   = "target"
+		componentName = "mysql"
+	)
+	scheme := runtime.NewScheme()
+	require.NoError(t, appsv1.AddToScheme(scheme))
+	cluster := &appsv1.Cluster{
+		ObjectMeta: metav1.ObjectMeta{Namespace: namespace, Name: clusterName},
+		Spec:       appsv1.ClusterSpec{Topology: "shared-nothing"},
+	}
+	component := &appsv1.Component{
+		ObjectMeta: metav1.ObjectMeta{
+			Namespace: namespace,
+			Name:      constant.GenerateClusterComponentName(clusterName, componentName),
+		},
+		Spec: appsv1.ComponentSpec{
+			ServiceVersion: "3.3.2",
+			Instances: []appsv1.InstanceTemplate{
+				{Name: "canary", ServiceVersion: "3.4.0"},
+				{Name: "inherited"},
+			},
+		},
+	}
+	cli := fake.NewClientBuilder().WithScheme(scheme).WithObjects(cluster, component).Build()
+	manager := &RestoreManager{Restore: &dpv1alpha1.Restore{}}
+	reqCtx := intctrlutil.RequestCtx{Ctx: context.Background()}
+
+	tests := []struct {
+		name             string
+		instanceTemplate string
+		wantVersion      string
+	}{
+		{name: "component default", wantVersion: "3.3.2"},
+		{name: "instance override", instanceTemplate: "canary", wantVersion: "3.4.0"},
+		{name: "instance inherits component", instanceTemplate: "inherited", wantVersion: "3.3.2"},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			pod := &corev1.Pod{ObjectMeta: metav1.ObjectMeta{
+				Namespace: namespace,
+				Name:      "target-mysql-0",
+				Labels: map[string]string{
+					constant.AppInstanceLabelKey:           clusterName,
+					constant.KBAppComponentLabelKey:        componentName,
+					constant.KBAppInstanceTemplateLabelKey: tt.instanceTemplate,
+				},
+			}}
+
+			env, err := manager.postReadyTargetEnv(reqCtx, cli, pod)
+
+			require.NoError(t, err)
+			require.Equal(t, []corev1.EnvVar{
+				{Name: dptypes.DPTargetClusterTopology, Value: "shared-nothing"},
+				{Name: dptypes.DPTargetComponentServiceVersion, Value: tt.wantVersion},
+			}, env)
+		})
+	}
+
+	env, err := manager.postReadyTargetEnv(reqCtx, cli, &corev1.Pod{})
+	require.NoError(t, err)
+	require.Empty(t, env, "non-KubeBlocks target Pods must keep their existing behavior")
 }
 
 var _ = Describe("RestoreManager Test", func() {
